refactor(gantz_egui): extract the shared plot leaf renderer
Move the egui_plot rendering (plot_body/plot_channel), value bounds,
channel splitting and colour resolution out of the Plot node into
ui_tree::plot, parameterized by a PlotParams struct buildable from
either the Plot node weight or a gantz_ui plot element. The Plot
node's ui/view_ui delegate through Plot::params, and the interpreter
now renders the plot element through the same leaf: bound state
splits into channels, absent w/h fill the available space, and the
mode attr is ignored for drawing (it selects how the producing node
accumulates state).

The split_channels shape tests move with the code, and value_bounds
gains coverage it previously lacked.
